BigIntegerField for facebook_id
Causes a dependency on Django 1.2.

Newer Facebook IDs are too big for the regular integer field.

+++ django_facebook/models.py
@@ -10,7 +10,7 @@ class FacebookProfileModel(models.Model):
     the fields in.
     '''
     about_me = models.TextField(blank=True, null=True)
-    facebook_id = models.IntegerField(blank=True, null=True)
+    facebook_id = models.BigIntegerField(blank=True, null=True)
     facebook_name = models.CharField(max_length=255, blank=True, null=True)
     facebook_profile_url = models.TextField(blank=True, null=True)
     website_url = models.TextField(blank=True, null=True)
